requirements.txt
Office365-REST-Python-Client
pyyaml
feedgen
fastapi
markdown
uvicorn